Split api.ts into domain modules behind a compatibility barrel
frontend/src/lib/api.ts (828 lines) becomes frontend/src/lib/api/ -
core.ts (shared request plumbing) plus search.ts, planner.ts,
observations.ts, operator.ts, map.ts (one domain each), reassembled
by index.ts into the exact same api object shape and the same ~20
named wrapper exports. Every existing call site in the frontend keeps
working unchanged - @/lib/api resolves to @/lib/api/index.ts
automatically. See docs/superpowers/specs/2026-08-06-split-api-ts-into-domain-modules-design.md.
